organized the file trees. The files not included in the tree are either under modification or bbeing removed.
