+++ graphql/resolvers/drinkResolver/availableDrink.js
@@ -1,6 +1,6 @@
-const AvailableDrink = require('../../models/drinkSchemas/availableDrink')
-const RegisteredDrink = require('../../models/drinkSchemas/registeredDrink')
-const { transformDrink } = require('./merge')
+const AvailableDrink = require('../../../models/drinkSchemas/availableDrink')
+const RegisteredDrink = require('../../../models/drinkSchemas/registeredDrink')
+const { transformDrink } = require('../merge')
 
 module.exports = {
   /**
